Fix repeated substitution bug
Given &z = [y] and &y = [x],
"**z" could end up being "x" instead of "*y" in some circumstances.
Since z is substituted before y and the intermediate result is "*y",
which then gets interpreted as a command for y substitution.

To fix this, replacement code was completely rewritten to
find all variable locations in the template first and then
replace them all at once, instead of repeating
search-and-replace over the whole string for each variable.

An additional benefit is that the output string isn't
re-allocated over and over again, but is pre-reserved and
constructed in one pass.

The same could've been most trivially achieved with Regex crate,
but having no dependencies is something of a priority.