Fix replacement for variables with no variants
Even if a variable has no text variants, i.e. coming from empty [],
it should still get replaced, instead of getting kept as "*var"
in the result.

This is a corner case that could still be useful to get right,
e.g. for nested macro generation involving zero-length tuples
(unit) as one of the cases.